Create packages from duke classes
