+++ data/duke.txt
@@ -1,3 +1,1 @@
-[E] [ ] haha (at: hahaha)
-[D] [ ] hahah (by: Dec 01 2019)
-[T] [ ] haha
+[T] [ ] 4
