+++ src/main/java/duke/task/Todo.java
@@ -1,0 +1,13 @@
+package duke.task;
+
+import duke.Task;
+
+public class Todo extends Task {
+    public Todo(String item){
+        this.setItem(item);
+    }
+
+    public String getTask() {
+        return "[T] " + this.getStatusIcon() + this.getItem();
+    }
+}
